requirements.txt
altair==5.5.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
extra-streamlit-components==0.1.80
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.39.0
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
psutil==6.1.1
pwdlib==0.2.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.40
streamlit==1.45.0
streamlit-authenticator==0.4.2
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
